vala-terminal: Fix null pointer deref in MainWindow.update_title().
A change in VTE behaviour resulted in SIGSEGV on new tab creation.
When no term title is set VTE now returns null title. This situation was
not handled in update_title() and resulted in null pointer dereference.
The fix checks for null title from VTE and uses a default value in this
case.